updates(sonal cloud): fix bugs, code smell issues
