+++ src/components/pages/CompanyRoles/index.tsx
@@ -39,32 +39,16 @@ export default function CompanyRoles() {
     CommonService.getCompanyRoles((data: any) => {
       if (url.indexOf('companyrolesappprovider') > 1) {
         setCompanyRoles(data.appProvider)
-        if (data.appProvider.subNavigation) {
-          setLinkArray(data.appProvider.subNavigation)
-        } else {
-          setLinkArray(data.subNavigation)
-        }
+        setLinkArray(data.appProvider.subNavigation)
       } else if (url.indexOf('companyrolesserviceprovider') > 1) {
         setCompanyRoles(data.serviceProvider)
-        if (data.serviceProvider.subNavigation) {
-          setLinkArray(data.serviceProvider.subNavigation)
-        } else {
-          setLinkArray(data.subNavigation)
-        }
+        setLinkArray(data.serviceProvider.subNavigation)
       } else if (url.indexOf('companyrolesconfirmitybody') > 1) {
         setCompanyRoles(data.confirmity)
-        if (data.serviceProvider.subNavigation) {
-          setLinkArray(data.serviceProvider.subNavigation)
-        } else {
-          setLinkArray(data.subNavigation)
-        }
+        setLinkArray(data.confirmity.subNavigation)
       } else {
         setCompanyRoles(data.participant)
-        if (data.participant.subNavigation) {
-          setLinkArray(data.participant.subNavigation)
-        } else {
-          setLinkArray(data.subNavigation)
-        }
+        setLinkArray(data.participant.subNavigation)
       }
     })
   }, [url, language])
