+++ src/components/shared/templates/StaticTemplateResponsive/StaticTemplate.scss
@@ -555,3 +555,13 @@ strong:hover .tooltiptext {
 .imageVideoTextSideBySide.padding-bottom-20 {
   padding-bottom: 20px;
 }
+
+.subDescription {
+  display: block;
+  padding-top: 10px;
+  padding-bottom: 10px;
+}
+
+.padding-bottom-20 {
+  padding-bottom: 20px;
+}
